+++ lexicon/forms/azz.py
@@ -1,29 +1,19 @@
 from django import forms
-from django.conf import settings
-from django.db.models import Q
-from django.utils.translation import ugettext_lazy as _
 
-from query_builder.forms import (
-    QueryBuilderForm,
-    QueryBuilderGlobalFiltersForm,
-    QueryBuilderDatasetsForm,
-    QueryBuilderBaseFormset,
-)
+from query_builder.forms import QueryBuilderForm, QueryBuilderBaseFormset
 
 from mesolex.config import LANGUAGES
-from mesolex.utils import (
-    Language,
-    to_vln,
-)
+from mesolex.utils import Language, to_vln
 
 from lexicon.documents import LexicalEntryDocument
+from lexicon.forms.base import LexiconQueryBuilderGlobalFiltersForm
 from lexicon.transformations.nahuat_orthography import nahuat_orthography
 
 
 Azz = Language('azz')
 
 
-class LexicalSearchFilterForm(QueryBuilderForm):
+class AzzLexicalSearchFilterForm(QueryBuilderForm):
     FILTERABLE_FIELDS = Azz.filterable_fields
     FILTERABLE_FIELDS_DICT = Azz.filterable_fields_dict
     ELASTICSEARCH_FIELDS = Azz.elasticsearch_fields
@@ -42,23 +32,7 @@ class LexicalSearchFilterForm(QueryBuilderForm):
         ]
 
 
-class LexiconQueryBuilderGlobalFiltersForm(QueryBuilderGlobalFiltersForm):
-    only_with_sound = forms.BooleanField(required=False)
-
-    def clean_only_with_sound(self):
-        only_with_sound = self.cleaned_data['only_with_sound']
-        if only_with_sound:
-            return Q(media__isnull=(not only_with_sound))
-        return Q()
-
-
-class LexiconQueryBuilderDatasetsForm(QueryBuilderDatasetsForm):
-    dataset = forms.ChoiceField(
-        choices=[(l['code'], l['label']) for l in LANGUAGES.values()],
-        widget=forms.Select(attrs={'class': 'custom-select'}),
-    )
-
-class BaseLexiconQueryComposerFormset(QueryBuilderBaseFormset):
+class BaseAzzLexiconQueryComposerFormset(QueryBuilderBaseFormset):
     global_filters_class = LexiconQueryBuilderGlobalFiltersForm
 
     FILTERABLE_FIELDS = Azz.filterable_fields + Azz.elasticsearch_fields
@@ -70,7 +44,7 @@ class BaseLexiconQueryComposerFormset(QueryBuilderBaseFormset):
     }
 
 
-LexicalSearchFilterFormset = forms.formset_factory(
-    LexicalSearchFilterForm,
-    formset=BaseLexiconQueryComposerFormset,
+AzzLexicalSearchFilterFormset = forms.formset_factory(
+    AzzLexicalSearchFilterForm,
+    formset=BaseAzzLexiconQueryComposerFormset,
 )
